posts_service/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
aio_pika
python-dotenv
pyjwt
emoji
joblib
scikit-learn
fasttext
pandas
